fix delete files chat

+++ src/Controller/ChatController.php
@@ -430,9 +430,9 @@ class ChatController extends AbstractController
                     $f = explode("/", $image);
                     $filename = $f[count($f) - 1];
                     $file = "/var/www/vhosts/frikiradar.com/app.frikiradar.com/images/chat/" . $conversationId . "/" . $filename;
-                    try {
+                    if (file_exists($file)) {
                         unlink($file);
-                    } catch (Exception $ex) {
+                    } else {
                         // "No se ha podido borrar el archivo"
                     }
                 }
@@ -442,9 +442,9 @@ class ChatController extends AbstractController
                     $f = explode("/", $audio);
                     $filename = $f[count($f) - 1];
                     $file = "/var/www/vhosts/frikiradar.com/app.frikiradar.com/images/chat/" . $conversationId . "/" . $filename;
-                    try {
+                    if (file_exists($file)) {
                         unlink($file);
-                    } catch (Exception $ex) {
+                    } else {
                         // "No se ha podido borrar el archivo"
                     }
                 }
